pyproject.toml
[tool.poetry]
name = "corner-ticketing"
version = "0.1.0"
description = ""
authors = ["chaeyeon <gkscodus11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.2"
django-extensions = "^3.2.3"
black = "^24.4.2"
isort = "^5.13.2"
django-enumfields = "^2.1.1"
djangorestframework-simplejwt = "^5.3.1"
django-allauth = "^0.63.3"
dj-rest-auth = "^6.0.0"
requests = "^2.32.3"
cryptography = "^42.0.8"
drf-yasg = "^1.21.7"
psycopg2 = "^2.9.9"
django-redis = "^5.4.0"
django-crontab = "^0.7.1"
django-filter = "^24.2"
celery = {extras = ["redis"], version = "^5.4.0"}
django-silk = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length=120
exclude = '''
    \.git
    | venv
    | ^.*\b(migrations)\b.*$
'''

[tool.isort]
line_length=120
skip=[".gitignore", "migrations", "venv"]